Dynamexil/requirements.txt
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
python-engineio
cryptography==2.3
pyjwt==1.5.3
paho-mqtt==1.3.1
tendo==0.2.12
six==1.11.0
Werkzeug==0.14.1
eventlet==0.24.1
redis
pyserial
dynamixel-sdk
requests
simplejson
nano